Confidence method "Alternative 2"
An alternative set of minimum requirements does not specify which Categories indicators should be measured in but includes stricter requirements for the overall assessment regarding the number of heavy metal and organic indicators

At least 2 distinct Heavy metal indicators are included in the assessment, in any Category (Penalty=50%)
At least 3 distinct Organic indicators are included in the assessment, in any Category (Penalty=50%)

This is a slightly corrected version of the last commit. Unintentionally, the previous version counted only "Org" or "HM" when counting types. Now "heavy metal", "heavy", "organic", "ORGANIC", etc. are accepted.